[保母評論] 大修，主要是加入AddCommentDialogFragment 其它： 1.Rating資料型態由int改為float 2.保母 list Card layout的調整 3.[新增]評論時的速度改善 OPEN - #32: [UI/UX] 讓使用者可以更新資料

+++ src/tw/tasker/babysitter/presenter/adapter/MyInfoWindowAdapter.java
@@ -64,7 +64,7 @@ public class MyInfoWindowAdapter implements InfoWindowAdapter {
 		return null;
 	}
 	
-	private float getRatingValue(int totalRating, int totalComment) {
+	private float getRatingValue(float totalRating, int totalComment) {
 		float avgRating = 0.0f; 
 		
 		if (totalComment != 0) {
